test(harness): the direct-SQL helpers are their own file
WsExecErr took harness.go past the 500-line ceiling. Split where the file
actually parts: everything left builds a world through the product's own
doors, while these three reach past it to the estate — to seed a shape no
endpoint produces, to count rows a reader would not be shown, and to
assert that a statement the schema must refuse is refused.
